+++ semgrep/tests/unit/test_yaml_parsing.py
@@ -1,13 +1,18 @@
 import io
+
+import pytest
+from attr import validate
 from tempfile import NamedTemporaryFile
 from textwrap import dedent
 
 from ruamel.yaml import YAML
+from jsonschema.exceptions import ValidationError
 
 from semgrep.config_resolver import Config
 from semgrep.config_resolver import parse_config_string
 from semgrep.config_resolver import validate_single_rule
 from semgrep.constants import RULES_KEY
+from semgrep.error import InvalidRuleSchemaError
 
 
 def test_parse_taint_rules():
@@ -102,3 +107,103 @@ def test_default_yaml_type_safe():
     # Unsafe, executes the system call
     unsafe_yaml = YAML(typ="unsafe")
     assert unsafe_yaml.load(io.StringIO(s)) == 0
+
+
+def test_invalid_metavariable_regex():
+    rule = dedent(
+        """
+        rules:
+        - id: boto3-internal-network
+          patterns:
+          - pattern-inside: $MODULE.client(host=$HOST)
+          - metavariable-regex:
+              metavariable: $HOST
+              regex: '192.168\.\d{1,3}\.\d{1,3}'
+              metavariable: $MODULE
+              regex: (boto|boto3)
+          message: "Boto3 connection to internal network"
+          languages: [python]
+          severity: ERROR
+        """
+    )
+
+    with pytest.raises(InvalidRuleSchemaError):
+        parse_config_string("testfile", rule, None)
+
+
+def test_invalid_metavariable_comparison():
+    rule = dedent(
+        """
+        rules:
+        - id: boto3-internal-network
+          patterns:
+          - pattern-inside: $MODULE.client(host=$HOST, port=$PORT)
+          - metavariable-comparison:
+              metavariable: $PORT
+              comparison: $PORT > 9999
+              metavariable: $MODULE
+              comparison: '(server|servers)'
+          message: "Boto3 connection to internal network"
+          languages: [python]
+          severity: ERROR
+        """
+    )
+
+    with pytest.raises(InvalidRuleSchemaError):
+        parse_config_string("testfile", rule, None)
+
+
+def test_invalid_metavariable_comparison2():
+    rule = dedent(
+        """
+        rules:
+        - id: boto3-internal-network
+          patterns:
+          - pattern-inside: $MODULE.client(host=$HOST, port=$PORT)
+          - metavariable-comparison:
+              metavariable: $PORT
+              comparison: $PORT > 9999
+              metavariable: $MODULE
+              regex: '(server|servers)'
+          message: "Boto3 connection to internal network"
+          languages: [python]
+          severity: ERROR
+        """
+    )
+
+    with pytest.raises(InvalidRuleSchemaError):
+        parse_config_string("testfile", rule, None)
+
+
+def test_invalid_pattern_child():
+    rule = dedent("""
+        rules:
+        - id: blah
+          message: blah
+          severity: INFO
+          languages: [python]
+          patterns:
+          - pattern-either:
+            - pattern: $X == $Y
+            - pattern-not: $Z == $Z
+    """)
+
+    with pytest.raises(InvalidRuleSchemaError):
+        parse_config_string("testfile", rule, None)
+
+
+def test_invalid_rule_with_null():
+    rule = dedent("""
+        rules:
+        - id: blah
+          message: ~
+          severity: INFO
+          languages: [python]
+          patterns:
+          - pattern-either:
+            - pattern: $X == $Y
+            - pattern-not: $Z == $Z
+    """)
+
+    with pytest.raises(InvalidRuleSchemaError):
+        parse_config_string("testfile", rule, None)